pyproject.toml
[tool.poetry]
name = "derive"
version = "0.1.0"
description = "The Python Observability Toolkit"
authors = ["DeBankDeFi <sre@debank.com>"]

readme = "README.md"
license = "MIT"
homepage = "https://github.com/DeBankDeFi/derive"
repository = "https://github.com/DeBankDeFi/derive"
documentation = "https://github.com/DeBankDeFi/derive"

keywords = ["python", "observability", "opentelemetry"]

classifiers = [
    "Topic :: Software Development",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",

]
[tool.poetry.dependencies]
python = "^3.7,>=3.7.2"
configalchemy = "^0.5.5"
opentelemetry-sdk = "^1.11.1"
opentelemetry-api = "^1.11.1"

opentelemetry-sdk-extension-aws = { version = "^2.0", optional = true }
opentelemetry-propagator-aws-xray = { version = "^1.0", optional = true }
opentelemetry-exporter-otlp = { version = "^1.11", optional = true }
protobuf = { version = "^3.10", optional = true }
opentelemetry-instrumentation-requests = { version = "0.30b1", optional = true }

[tool.poetry.extras]
otlp-aws-xray = ["opentelemetry-sdk-extension-aws", "opentelemetry-propagator-aws-xray", "opentelemetry-exporter-otlp", "protobuf"]
requests = ["opentelemetry-instrumentation-requests"]

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pre-commit = "^2.20.0"
black = "^22.6.0"
mypy = "^0.961"
opentelemetry-sdk-extension-aws = "2.0.1"
opentelemetry-propagator-aws-xray = "1.0.1"
opentelemetry-exporter-otlp = "1.11.1"
protobuf = "^3.10"
coverage = "^6.4.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
files = "derive"
mypy_path = "derive"
ignore_missing_imports = true
namespace_packages = true
show_error_codes = true
install_types = true
non_interactive = true
strict = false
enable_error_code = [
    "ignore-without-code",
    "redundant-expr",
    "truthy-bool",
]